+++ meta.go
@@ -14,7 +14,7 @@ import (
 )
 
 var (
-	version = "3.0.0"
+	version = "3.0.3"
 	commit  = "unknown"
 	builtAt = "unknown"
 )